Consumer: revamp message tracking
Essentially, revisit the internals of the consumer application to remove old and messy code.
The new supervision structure brought forward in #6 was the first step,
and now the actual implementation of AMQPDataConsumer, MessageTracker and DataUpdater is simplified.

At a first glance, the most relevant change is that it is the MessageTracker
(and not the AMQPDataConsumer) tasked with ack-/reject-ing a message, once processing is done.
This allows to remove a level of indirection and communication between processes,
making the system more reliable and fault-tolerant, because the application state is less distributed.

Messages are received by an AMQPDataConsumer, then forwarded via cast to the MessageTracker identified
by the sharding key, together with a reference to the AMQP Channel on which to ack.
The MessageTracker then forwards via cast the Message to the DataUpdater identified by the same
sharding key, which replies via a call when the message has finished processing.Finally,
the MessageTracker uses the Channel reference to ack the message (in a safe way).
This flow allows the MessageTracker to maintain the ordering inariant, which is internally done by having
a FIFO queue in the state and making sure that only the message in the head is processed.

For fault tolerance, the following supervisions are put in place:
- Each AMQPDataConsumer is linked with the Channel it receives on
- Every DataConsumer monitors each MessageTracker it firwards messages to
- Each MessageTracker monitors the Channel whose message is forwarded
- Each MessageTracker monitors the DataUpdater identified by the same sharding key

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>