fix(events_producer): reconnection
make reconnection work by correctly handling the messages

one big change of the current refactor is that in case of an error during startup,
the producer will enter the reconnection retry loop instead of dying.

this way we handle the error ourself instead of leaving the job to the supervisor,
which is preferable as we already have the reconnection logic.

also sets up the groundwork for using the AMQPConnection behavior

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>