refactor: use Horde's Registry and DynamicSupervisor
Replace Kernel's Registry and DynamicSupervisor modules
with Horde's distributed versions.

We still need to set up clustering with libcluster next.

A fork of horde is used instead of the version from hex because
of a bug where `extra_arguments` is not passed as a parameter to the
processes in a DynamicSupervisor

Signed-off-by: Francesco Noacco <francesco.noacco@secomind.com>