Consumer: avoid duplicate processes in distributed setting
When using Horde in a distributed cluster, it might happen
that two processes for the same registry value are spawned.
While Horde takes care of shutting down the "wrong" one,
we have to take care that it is not restarted by the supervisor.
See https://hexdocs.pm/horde/eventual_consistency.html#horde-registry-merge-conflict.

So:
- Make the AMQPConsumerSupervisor exit with `:normal` when duplicated,
  that way it won't be restarted
- Make DataUpdater and MessageTraker processes temporary so that they
  will be restarted only when requested.

Signed-off-by: Arnaldo Cesco <arnaldo.cesco@secomind.com>